+++ src/main/kotlin/com/example/demo/model/UsdPrice.kt
@@ -1,4 +1,4 @@
 package com.example.demo.model
 
-data class UsdPrice (val d: String, val v: Float)
+data class UsdPrice (val d: String, val v: Double)
